Fix clang warning

+++ src/save.c
@@ -119,7 +119,7 @@ void hangup(int sig_unused) {
 	if (!program_state.done_hup++) {
 		if (program_state.something_worth_saving) {
 			// AIS: record levels on which there were hangups
-			if (ledger_no(&u.uz) >= 0 && ledger_no(&u.uz) < MAXLINFO) {
+			if (ledger_no(&u.uz) >= 0 && (int)ledger_no(&u.uz) < MAXLINFO) {
 				level_info[ledger_no(&u.uz)].flags |= HANGUP_HERE;
 			}
 			dosave0();
